Some files reorganization
 - Renamed ParserCommand to ParserCLI to avoid confusion with "…Commands" named files used to declare Commander "Commands" instances
 - Split Config & Config.Entry in two separate files
 - Split Config tests in two (reading & linting)
 - Fix formatting / documentation / access modifiers (some stuff made private)
